fix(types): align the mypy strict gate with the current toolchain (#240)
make typecheck (mypy 1.20 per dev extra >=1.11) surfaced drift the
historical 1.1x run never saw:

- types-PyYAML added to the dev extra — mypy --strict needs the stubs,
  the historical dev env carried them ad hoc
- training/ excluded from strict checking (ADR-0021 anti-scope: never
  ships, not runtime code) — mypy followed mnemos.train_entry into it
- api.get_memory: the sloppy 'request: Request = None' default became
  an error under current fastapi typing; Request is always injected by
  FastAPI, so the parameter is now required and ordered before defaults
- _operator_context accepts Request | None with a defensive None branch
  (no injected request = no admitted session = not operator context)
- train_entry/refine: int() narrowing on Any-returning calls

Gates: ruff check + format clean, mypy --strict 86 files clean,
full suite 2362 passed.

+++ src/mnemos/pipeline/refine.py
@@ -113,7 +113,7 @@ def _swap_key(cluster_context: str, processed_content: str) -> str:
 
 def _retry_backoff_sec(attempt: int) -> int:
     """Exponential backoff for lane (a), capped (DLQ convention)."""
-    return min(REFINE_BACKOFF_BASE_SEC * (2 ** (attempt - 1)), REFINE_BACKOFF_CAP_SEC)
+    return int(min(REFINE_BACKOFF_BASE_SEC * (2 ** (attempt - 1)), REFINE_BACKOFF_CAP_SEC))
 
 
 def _revision_hash(text: str) -> str:
